--- src/components/start-airdrop/PickRoles/index.tsx
@@ -1,3 +1,0 @@
-import PickRoles from "./PickRoles"
-
-export default PickRoles
