+++ src/components/start-airdrop/UploadNFTs/components/RoleCard/components/TraitInput.tsx
