+++ src/components/start-airdrop/UploadNFTs/index.tsx
@@ -1,0 +1,3 @@
+import UploadNFTs from "./UploadNFTs"
+
+export default UploadNFTs
