+++ src/pages/api/get-signature/claim.ts
@@ -2,7 +2,7 @@ import { defaultAbiCoder } from "@ethersproject/abi"
 import { arrayify } from "@ethersproject/bytes"
 import { keccak256 } from "@ethersproject/keccak256"
 import { Wallet } from "@ethersproject/wallet"
-import { fetchRoles } from "components/start-airdrop/PickRoles/hooks/useRoles"
+import { fetchRoles } from "components/start-airdrop/UploadNFTs/hooks/useRoles"
 import { fetchUserRoles } from "components/[drop]/ClaimCard/hooks/useUserRoles"
 import { Chains } from "connectors"
 import { AirdropAddresses } from "contracts"
